+++ src/fresco/containers/geometricShapes/utils/Point.java
@@ -4,32 +4,32 @@ import fresco.containers.geometricShapes.Circle;
 import fresco.containers.geometricShapes.Line;
 
 public class Point {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
     public Point() {
         this.x = 0;
         this.y = 0;
     }
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 
@@ -37,20 +37,20 @@ public class Point {
         return Math.sqrt(Math.pow(b.getX() - a.getX(),2) + Math.pow(b.getY() - a.getY(),2));
     }
 
-    public void translate(double x_translation, double y_translation){
+    public void translate(int x_translation, int y_translation){
         this.setX(x + x_translation);
         this.setY(y + y_translation);
     }
 
     public void centralSymmetry(Point o){
-        double x_distance = o.getX() - x;
-        double y_distance = o.getY() - y;
+        int x_distance = o.getX() - x;
+        int y_distance = o.getY() - y;
         translate(x_distance*2, y_distance*2);
     }
 
     public void axialSymmetry(Line l){
-        double x_distance = (l.getPointA().getX() - x) + (l.getPointB().getX() - x);
-        double y_distance = (l.getPointA().getY() - y) + (l.getPointB().getY() - y);
+        int x_distance = (l.getPointA().getX() - x) + (l.getPointB().getX() - x);
+        int y_distance = (l.getPointA().getY() - y) + (l.getPointB().getY() - y);
         translate(x_distance, y_distance);
     }
